Move query result envelope helpers into partial
